Adicionando forma de pagamento com paypal

+++ database/migrations/2022_11_02_210209_create_order_table.php
@@ -17,7 +17,7 @@ class CreateOrderTable extends Migration
             $table->id();
             $table->unsignedBigInteger('id_cart');
             $table->unsignedBigInteger('id_user');
-            $table->unsignedTinyInteger('total_value');
+            $table->float('total_value');
             $table->timestamps();
         });
